Simplify test logic
If we need a browser, we'll need a static server for the page so we can
merge these into one single macro, removing the need for each test to
manually setup the before and after test hooks.
