+++ tests/loading-screenshot.js
@@ -1,18 +1,14 @@
 /* global XRPackage */
 const test = require('ava');
 
-const withPage = require('./utils/_withPage');
-const {testBeforeHook, testAfterHook} = require('./utils/_testHelpers');
+const withPageAndStaticServer = require('./utils/_withPageAndStaticServer');
 
-test.before(testBeforeHook);
-test.after.always(testAfterHook);
-
-test.serial('load screenshot of baked xrpk', withPage, async (t, page) => {
+test('load screenshot of baked xrpk', withPageAndStaticServer, async (t, page) => {
   const response = await page.evaluate(pageFunction, `${t.context.staticUrl}/assets/baked-xrpk.wbn`);
   t.true(response && response.startsWith('<img src="blob:http://localhost:'));
 });
 
-test.serial('load screenshot of unbaked xrpk', withPage, async (t, page) => {
+test('load screenshot of unbaked xrpk', withPageAndStaticServer, async (t, page) => {
   const response = await page.evaluate(pageFunction, `${t.context.staticUrl}/assets/unbaked-xrpk.wbn`);
   t.falsy(response);
 });
